+++ src/core/acroform/PDFAcroNonTerminal.ts
@@ -2,6 +2,7 @@ import PDFDict from 'src/core/objects/PDFDict';
 import PDFRef from 'src/core/objects/PDFRef';
 import PDFContext from 'src/core/PDFContext';
 import PDFAcroField from 'src/core/acroform/PDFAcroField';
+import PDFName from '../objects/PDFName';
 
 class PDFAcroNonTerminal extends PDFAcroField {
   static fromDict = (dict: PDFDict, ref: PDFRef) =>
@@ -16,6 +17,17 @@ class PDFAcroNonTerminal extends PDFAcroField {
   addField(field: PDFRef) {
     const { Kids } = this.normalizedEntries();
     Kids?.push(field);
+  }
+
+  normalizedEntries() {
+    let Kids = this.Kids();
+
+    if (!Kids) {
+      Kids = this.dict.context.obj([]);
+      this.dict.set(PDFName.of('Kids'), Kids);
+    }
+
+    return { Kids };
   }
 }
