+++ src/utils/validators.ts
@@ -1,6 +1,6 @@
 /* tslint:disable:ban-types */
 
-import { values as objectValues } from 'src/utils';
+import { values as objectValues } from 'src/utils/objects';
 
 export const backtick = (val: any) => `\`${val}\``;
 export const singleQuote = (val: any) => `'${val}'`;
